+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ## 6.7.3 (unreleased)
 
 - GitHub Importer: fix import error with issues associated to a closed milestone.
+- Trello Importer: fix import error with attachemts without owner.
+- Trello Importer: fix import error when attachemt name ends with '/'.
 
 ## 6.7.2 (2024-02-16)